split files with multiple classes to separate files, split to folders, renames
