--- src/Microsoft.TemplateEngine.Cli/Commands/InstallCommand.cs
@@ -1,117 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-#nullable enable
-
-using System.CommandLine;
-using System.CommandLine.Invocation;
-using System.CommandLine.Parsing;
-using Microsoft.TemplateEngine.Abstractions;
-using Microsoft.TemplateEngine.Edge.Settings;
-
-namespace Microsoft.TemplateEngine.Cli.Commands
-{
-    internal class InstallCommand : BaseInstallCommand
-    {
-        public InstallCommand(
-                NewCommand parentCommand,
-                ITemplateEngineHost host,
-                ITelemetryLogger logger,
-                NewCommandCallbacks callbacks)
-            : base(parentCommand, host, logger, callbacks, "install")
-        {
-            parentCommand.AddNoLegacyUsageValidators(this);
-        }
-    }
-
-    internal class LegacyInstallCommand : BaseInstallCommand
-    {
-        public LegacyInstallCommand(NewCommand parentCommand, ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks)
-            : base(parentCommand, host, logger, callbacks, "--install")
-        {
-            this.IsHidden = true;
-            this.AddAlias("-i");
-
-            parentCommand.AddNoLegacyUsageValidators(this, except: new Option[] { InteractiveOption, AddSourceOption });
-        }
-
-        internal override Option<bool> InteractiveOption => ParentCommand.InteractiveOption;
-
-        internal override Option<IReadOnlyList<string>> AddSourceOption => ParentCommand.AddSourceOption;
-    }
-
-    internal abstract class BaseInstallCommand : BaseCommand<InstallCommandArgs>
-    {
-        internal BaseInstallCommand(
-            NewCommand parentCommand,
-            ITemplateEngineHost host,
-            ITelemetryLogger logger,
-            NewCommandCallbacks callbacks,
-            string commandName)
-            : base(host, logger, callbacks, commandName, SymbolStrings.Command_Install_Description)
-        {
-            ParentCommand = parentCommand;
-            this.AddArgument(NameArgument);
-            this.AddOption(InteractiveOption);
-            this.AddOption(AddSourceOption);
-        }
-
-        internal Argument<IReadOnlyList<string>> NameArgument { get; } = new("package")
-        {
-            Description = SymbolStrings.Command_Install_Argument_Package,
-            Arity = new ArgumentArity(1, 99)
-        };
-
-        internal virtual Option<bool> InteractiveOption { get; } = SharedOptionsFactory.CreateInteractiveOption();
-
-        internal virtual Option<IReadOnlyList<string>> AddSourceOption { get; } = SharedOptionsFactory.CreateAddSourceOption();
-
-        protected NewCommand ParentCommand { get; }
-
-        protected override async Task<NewCommandStatus> ExecuteAsync(InstallCommandArgs args, IEngineEnvironmentSettings environmentSettings, InvocationContext context)
-        {
-            using TemplatePackageManager templatePackageManager = new TemplatePackageManager(environmentSettings);
-            TemplatePackageCoordinator templatePackageCoordinator = new TemplatePackageCoordinator(
-                TelemetryLogger,
-                environmentSettings,
-                templatePackageManager);
-
-            //we need to await, otherwise templatePackageManager will be disposed.
-            return await templatePackageCoordinator.EnterInstallFlowAsync(args, context.GetCancellationToken()).ConfigureAwait(false);
-        }
-
-        protected override InstallCommandArgs ParseContext(ParseResult parseResult)
-        {
-            return new InstallCommandArgs(this, parseResult);
-        }
-    }
-
-    internal class InstallCommandArgs : GlobalArgs
-    {
-        public InstallCommandArgs(BaseInstallCommand installCommand, ParseResult parseResult) : base(installCommand, parseResult)
-        {
-            TemplatePackages = parseResult.GetValueForArgument(installCommand.NameArgument)
-                ?? throw new ArgumentException($"{nameof(parseResult)} should contain at least one argument for {nameof(installCommand.NameArgument)}", nameof(parseResult));
-
-            //workaround for --install source1 --install source2 case
-            if (installCommand is LegacyInstallCommand && installCommand.Aliases.Any(alias => TemplatePackages.Contains(alias)))
-            {
-                TemplatePackages = TemplatePackages.Where(package => !installCommand.Aliases.Contains(package)).ToList();
-            }
-
-            if (!TemplatePackages.Any())
-            {
-                throw new ArgumentException($"{nameof(parseResult)} should contain at least one argument for {nameof(installCommand.NameArgument)}", nameof(parseResult));
-            }
-
-            Interactive = parseResult.GetValueForOption(installCommand.InteractiveOption);
-            AdditionalSources = parseResult.GetValueForOption(installCommand.AddSourceOption);
-        }
-
-        public IReadOnlyList<string> TemplatePackages { get; }
-
-        public bool Interactive { get; }
-
-        public IReadOnlyList<string>? AdditionalSources { get; }
-    }
-}
