+++ src/Microsoft.TemplateEngine.Cli/Commands/alias/BaseAliasAddCommand.cs
@@ -9,23 +9,6 @@ using Microsoft.TemplateEngine.Abstractions;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal class AliasAddCommand : BaseAliasAddCommand
-    {
-        internal AliasAddCommand(ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks) : base(host, logger, callbacks, "add")
-        {
-            IsHidden = true;
-        }
-    }
-
-    internal class LegacyAliasAddCommand : BaseAliasAddCommand
-    {
-        internal LegacyAliasAddCommand(ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks) : base(host, logger, callbacks, "--alias")
-        {
-            AddAlias("-a");
-            IsHidden = true;
-        }
-    }
-
     internal class BaseAliasAddCommand : BaseCommand<AliasAddCommandArgs>
     {
         internal BaseAliasAddCommand(
@@ -38,12 +21,5 @@ namespace Microsoft.TemplateEngine.Cli.Commands
         protected override Task<NewCommandStatus> ExecuteAsync(AliasAddCommandArgs args, IEngineEnvironmentSettings environmentSettings, InvocationContext context) => throw new NotImplementedException();
 
         protected override AliasAddCommandArgs ParseContext(ParseResult parseResult) => new(this, parseResult);
-    }
-
-    internal class AliasAddCommandArgs : GlobalArgs
-    {
-        public AliasAddCommandArgs(BaseAliasAddCommand command, ParseResult parseResult) : base(command, parseResult)
-        {
-        }
     }
 }
