+++ src/Microsoft.TemplateEngine.Cli/Commands/alias/BaseAliasShowCommand.cs
@@ -9,22 +9,6 @@ using Microsoft.TemplateEngine.Abstractions;
 
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
-    internal class AliasShowCommand : BaseAliasShowCommand
-    {
-        internal AliasShowCommand(ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks) : base(host, logger, callbacks, "show")
-        {
-            IsHidden = true;
-        }
-    }
-
-    internal class LegacyAliasShowCommand : BaseAliasShowCommand
-    {
-        internal LegacyAliasShowCommand(ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks) : base(host, logger, callbacks, "--show-alias")
-        {
-            IsHidden = true;
-        }
-    }
-
     internal class BaseAliasShowCommand : BaseCommand<AliasShowCommandArgs>
     {
         internal BaseAliasShowCommand(
@@ -37,12 +21,5 @@ namespace Microsoft.TemplateEngine.Cli.Commands
         protected override Task<NewCommandStatus> ExecuteAsync(AliasShowCommandArgs args, IEngineEnvironmentSettings environmentSettings, InvocationContext context) => throw new NotImplementedException();
 
         protected override AliasShowCommandArgs ParseContext(ParseResult parseResult) => new(this, parseResult);
-    }
-
-    internal class AliasShowCommandArgs : GlobalArgs
-    {
-        public AliasShowCommandArgs(BaseAliasShowCommand command, ParseResult parseResult) : base(command, parseResult)
-        {
-        }
     }
 }
