+++ src/Microsoft.TemplateEngine.Cli/Commands/alias/LegacyAliasAddCommand.cs
@@ -1,0 +1,18 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+using Microsoft.TemplateEngine.Abstractions;
+
+namespace Microsoft.TemplateEngine.Cli.Commands
+{
+    internal class LegacyAliasAddCommand : BaseAliasAddCommand
+    {
+        internal LegacyAliasAddCommand(ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks) : base(host, logger, callbacks, "--alias")
+        {
+            AddAlias("-a");
+            IsHidden = true;
+        }
+    }
+}
