+++ src/Microsoft.TemplateEngine.Cli/Commands/alias/LegacyAliasShowCommand.cs
@@ -1,0 +1,17 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+using Microsoft.TemplateEngine.Abstractions;
+
+namespace Microsoft.TemplateEngine.Cli.Commands
+{
+    internal class LegacyAliasShowCommand : BaseAliasShowCommand
+    {
+        internal LegacyAliasShowCommand(ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks) : base(host, logger, callbacks, "--show-alias")
+        {
+            IsHidden = true;
+        }
+    }
+}
