+++ src/Microsoft.TemplateEngine.Cli/Commands/create/InvalidTemplateOptionResult.cs
@@ -10,50 +10,6 @@ using Microsoft.TemplateEngine.Abstractions;
 namespace Microsoft.TemplateEngine.Cli.Commands
 {
     /// <summary>
-    /// The class represents the information about the template option used when executing the command.
-    /// </summary>
-    internal class TemplateOptionResult
-    {
-        internal TemplateOptionResult(
-          TemplateOption? templateOption,
-          string inputFormat,
-          string? specifiedValue)
-        {
-            TemplateOption = templateOption;
-            InputFormat = inputFormat;
-            SpecifiedValue = specifiedValue;
-        }
-
-        /// <summary>
-        /// the alias used in CLI for parameter.
-        /// </summary>
-        internal string InputFormat { get; }
-
-        /// <summary>
-        /// The value specified for the parameter in CLI.
-        /// </summary>
-        internal string? SpecifiedValue { get; }
-
-        internal TemplateOption? TemplateOption { get; }
-
-        internal static TemplateOptionResult? FromParseResult(TemplateOption option, ParseResult parseResult)
-        {
-            OptionResult? optionResult = parseResult.FindResultFor(option.Option);
-
-            if (optionResult == null)
-            {
-                //option is not specified
-                return null;
-            }
-
-            return new TemplateOptionResult(
-                    option,
-                    optionResult.Token.Value ?? string.Empty,
-                    optionResult.GetValueOrDefault<string>());
-        }
-    }
-
-    /// <summary>
     /// The class represents the information about the invalid template option used when executing the command.
     /// </summary>
     internal class InvalidTemplateOptionResult : TemplateOptionResult, IEquatable<InvalidTemplateOptionResult>
