+++ src/Microsoft.TemplateEngine.Cli/Commands/create/TemplateResult.cs
