+++ src/Microsoft.TemplateEngine.Cli/Commands/install/LegacyInstallCommand.cs
@@ -1,0 +1,26 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+using System.CommandLine;
+using Microsoft.TemplateEngine.Abstractions;
+
+namespace Microsoft.TemplateEngine.Cli.Commands
+{
+    internal class LegacyInstallCommand : BaseInstallCommand
+    {
+        public LegacyInstallCommand(NewCommand parentCommand, ITemplateEngineHost host, ITelemetryLogger logger, NewCommandCallbacks callbacks)
+            : base(parentCommand, host, logger, callbacks, "--install")
+        {
+            this.IsHidden = true;
+            this.AddAlias("-i");
+
+            parentCommand.AddNoLegacyUsageValidators(this, except: new Option[] { InteractiveOption, AddSourceOption });
+        }
+
+        internal override Option<bool> InteractiveOption => ParentCommand.InteractiveOption;
+
+        internal override Option<IReadOnlyList<string>> AddSourceOption => ParentCommand.AddSourceOption;
+    }
+}
