+++ src/Microsoft.TemplateEngine.Cli/Commands/search/SearchCommand.cs
@@ -1,0 +1,22 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+using Microsoft.TemplateEngine.Abstractions;
+
+namespace Microsoft.TemplateEngine.Cli.Commands
+{
+    internal class SearchCommand : BaseSearchCommand
+    {
+        public SearchCommand(
+                NewCommand parentCommand,
+                ITemplateEngineHost host,
+                ITelemetryLogger logger,
+                NewCommandCallbacks callbacks)
+            : base(parentCommand, host, logger, callbacks, "search")
+        {
+            parentCommand.AddNoLegacyUsageValidators(this);
+        }
+    }
+}
