+++ src/Microsoft.TemplateEngine.Cli/Commands/uninstall/LegacyUninstallCommand.cs
@@ -1,0 +1,25 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+using Microsoft.TemplateEngine.Abstractions;
+
+namespace Microsoft.TemplateEngine.Cli.Commands
+{
+    internal class LegacyUninstallCommand : BaseUninstallCommand
+    {
+        public LegacyUninstallCommand(
+            NewCommand parentCommand,
+            ITemplateEngineHost host,
+            ITelemetryLogger logger,
+            NewCommandCallbacks callbacks)
+            : base(host, logger, callbacks, "--uninstall")
+        {
+            this.IsHidden = true;
+            this.AddAlias("-u");
+
+            parentCommand.AddNoLegacyUsageValidators(this);
+        }
+    }
+}
