+++ src/Microsoft.TemplateEngine.Cli/Commands/update/BaseUpdateCommand.cs
@@ -1,0 +1,50 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+using System.CommandLine;
+using System.CommandLine.Invocation;
+using System.CommandLine.Parsing;
+using Microsoft.TemplateEngine.Abstractions;
+using Microsoft.TemplateEngine.Edge.Settings;
+
+namespace Microsoft.TemplateEngine.Cli.Commands
+{
+    internal class BaseUpdateCommand : BaseCommand<UpdateCommandArgs>
+    {
+        internal BaseUpdateCommand(
+            NewCommand parentCommand,
+            ITemplateEngineHost host,
+            ITelemetryLogger logger,
+            NewCommandCallbacks callbacks,
+            string commandName,
+            string description)
+            : base(host, logger, callbacks, commandName, description)
+        {
+            ParentCommand = parentCommand;
+            this.AddOption(InteractiveOption);
+            this.AddOption(AddSourceOption);
+        }
+
+        internal virtual Option<bool> InteractiveOption { get; } = SharedOptionsFactory.CreateInteractiveOption();
+
+        internal virtual Option<IReadOnlyList<string>> AddSourceOption { get; } = SharedOptionsFactory.CreateAddSourceOption();
+
+        protected NewCommand ParentCommand { get; }
+
+        protected override async Task<NewCommandStatus> ExecuteAsync(UpdateCommandArgs args, IEngineEnvironmentSettings environmentSettings, InvocationContext context)
+        {
+            using TemplatePackageManager templatePackageManager = new TemplatePackageManager(environmentSettings);
+            TemplatePackageCoordinator templatePackageCoordinator = new TemplatePackageCoordinator(
+                TelemetryLogger,
+                environmentSettings,
+                templatePackageManager);
+
+            //we need to await, otherwise templatePackageManager will be disposed.
+            return await templatePackageCoordinator.EnterUpdateFlowAsync(args, context.GetCancellationToken()).ConfigureAwait(false);
+        }
+
+        protected override UpdateCommandArgs ParseContext(ParseResult parseResult) => new(this, parseResult);
+    }
+}
