+++ src/Microsoft.TemplateEngine.Cli/PublicAPI.Shipped.txt
@@ -1,6 +1,4 @@
 ﻿#nullable enable
-Microsoft.TemplateEngine.Cli.CommandParserException
-Microsoft.TemplateEngine.Cli.CommandParserException.Argument.get -> string!
 Microsoft.TemplateEngine.Cli.Components
 Microsoft.TemplateEngine.Cli.HostSpecificDataLoader
 Microsoft.TemplateEngine.Cli.HostSpecificDataLoader.HostSpecificDataLoader(Microsoft.TemplateEngine.Abstractions.IEngineEnvironmentSettings! engineEnvironment) -> void
