+++ test/Microsoft.TemplateEngine.Cli.UnitTests/ParserTests/InstantiateTests.Subcommand.cs
@@ -82,7 +82,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.ParserTests
             TemplateCommand templateCommand = new TemplateCommand(instantiateCommand, settings, packageManager, templateGroup, templateGroup.Templates.Single());
             Parser parser = ParserFactory.CreateParser(templateCommand);
             ParseResult templateParseResult = parser.Parse(args.RemainingArguments ?? Array.Empty<string>());
-            var templateArgs = new TemplateArgs(templateCommand, templateParseResult);
+            var templateArgs = new TemplateCommandArgs(templateCommand, templateParseResult);
 
             Assert.Equal(expectedValue, templateArgs.Name);
         }
@@ -115,7 +115,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.ParserTests
             TemplateCommand templateCommand = new TemplateCommand(instantiateCommand, settings, packageManager, templateGroup, templateGroup.Templates.Single());
             Parser parser = ParserFactory.CreateParser(templateCommand);
             ParseResult templateParseResult = parser.Parse(args.RemainingArguments ?? Array.Empty<string>());
-            var templateArgs = new TemplateArgs(templateCommand, templateParseResult);
+            var templateArgs = new TemplateCommandArgs(templateCommand, templateParseResult);
 
             if (string.IsNullOrWhiteSpace(expectedValue))
             {
@@ -150,7 +150,7 @@ namespace Microsoft.TemplateEngine.Cli.UnitTests.ParserTests
             TemplateCommand templateCommand = new TemplateCommand(instantiateCommand, settings, packageManager, templateGroup, templateGroup.Templates.Single());
             Parser parser = ParserFactory.CreateParser(templateCommand);
             ParseResult templateParseResult = parser.Parse(args.RemainingArguments ?? Array.Empty<string>());
-            var templateArgs = new TemplateArgs(templateCommand, templateParseResult);
+            var templateArgs = new TemplateCommandArgs(templateCommand, templateParseResult);
 
             if (string.IsNullOrWhiteSpace(expectedValue))
             {
