+++ test/Microsoft.TemplateEngine.Cli.UnitTests/ParserTests/TemplateCommandTests.cs
@@ -7,7 +7,6 @@ using System.CommandLine;
 using FakeItEasy;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Cli.Commands;
-using Microsoft.TemplateEngine.Cli.Commands.Exceptions;
 using Microsoft.TemplateEngine.Edge;
 using Microsoft.TemplateEngine.Edge.Settings;
 using Microsoft.TemplateEngine.Mocks;
